requirements.txt
beautifulsoup4
grobid-client-python
lxml
pandas
pyarrow
python-dotenv
elasticsearch
tqdm